Struct loggerv::Logger

pub struct Logger { /* fields omitted */ }

Methods

impl Logger

pub fn new() -> Logger

Creates a new instance of the verbosity-based logger.

The default level is WARN. Color is enabled if the parent application or library is running from a terminal, i.e. running a tty. The default separator is the ": " string. The default output format is module path: message. The following default colors are used:

Level 	Color

Error 	Bright Red

Warn 	Bright Yellow

Info 	Bright Green

Debug 	Light Grey

Trace 	Grey

pub fn error_color(self, c: Colour) -> Self

Sets the output color for the ERROR level. The default is bright red.

pub fn warn_color(self, c: Colour) -> Self

Sets the output color for the WARN level. The default is bright yellow.

pub fn info_color(self, c: Colour) -> Self

Sets the output color for the INFO level. The default is bright green.

pub fn debug_color(self, c: Colour) -> Self

Sets the output color for the DEBUG level. The default is light grey.

pub fn trace_color(self, c: Colour) -> Self

Sets the output color for the TRACE level. The default is grey.

pub fn separator(self, s: &str) -> Self

Sets the separator string.

The separator is the string between the "tag" and the message that make up a log statement. The tag will be colorized if enabled, while the message will not. The default is :.

If the level, line numbers, and module path are all not included in the log statement, then the separator is changed to the empty string to avoid printing a lone string or character before each message portion of the log statement.

pub fn colors(self, c: bool) -> Self

Enables or disables colorizing the output.

If the logger is not used in a terminal, then the output is not colorized regardless of this value.

pub fn no_colors(self) -> Self

Disables colorizing the output.

The default is to colorize the output unless stdout and stderr are redirected or piped, i.e. not a tty.

pub fn line_numbers(self, i: bool) -> Self

Enables or disables including line numbers in the "tag" portion of the log statement. The tag is the text to the left of the separator.

pub fn level(self, i: bool) -> Self

Enables or disables including the level in the log statement's tag portion. The tag of the log statement is the text to the left of the separator.

If the level and the module path are both inculded, then the module path is surrounded by square brackets.

pub fn max_level(self, l: LogLevel) -> Self

Explicitly sets the log level instead of through a verbosity.

pub fn module_path(self, i: bool) -> Self

Enables or disables including the module path in the "tag" portion of the log statement.

The tag is the text to the left of the separator. The default is to include the module path. Ifthe level is also included, the module path is surrounded by square brackets.

pub fn no_module_path(self) -> Self

Disables the module path in the "tag" portion of the log statement.

The tag is the text to the left of the separator. The default is to include the module path.

pub fn verbosity(self, v: u64) -> Self

Converts the verbosity to a log level.

A verbosity of zero (0) is the default, which means ERROR and WARN log statements are printed to stderr. No other log statements are printed on any of the standard streams (stdout or stderr). As the verbosity is increased, the log level is increased and more log statements will be printed to stdout. A verbosity of 1 will print INFO log statements to stdout in addition to ERROR and WARN. A verbosity of 2 will print INFO and DEBUG log statements to stdout. A verbosity of 3 or higher will print INFO, DEBUG, and TRACE log statements to stdout with ERROR and WARN statements printed to stderr.

pub fn init(self) -> Result<(), SetLoggerError>

Initializes the logger.

This also consumes the logger. It cannot be further modified after initialization.

If the tag will be empty because the level, line numbers, and module path were all disabled, then the separator is changed to the empty string to avoid writing a long character in front of each message for each log statement.